Project finished

+++ src/main/java/com/cursojava/curso/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.cursojava.curso.controllers;
 
 import com.cursojava.curso.dao.UserDao;
 import com.cursojava.curso.models.User;
+import com.cursojava.curso.utils.JWTUtil;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,8 @@ import java.util.List;
 public class UserController {
     @Autowired
     private UserDao userDao;
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @RequestMapping(value = "api/user/{id}",method = RequestMethod.GET)
     public User getUser(@PathVariable Long id){
@@ -23,40 +26,36 @@ public class UserController {
 
     //Get all users
     @RequestMapping(value = "api/users",method = RequestMethod.GET)
-    public List<User> getUsers(){
+    public List<User> getUsers(@RequestHeader (value="Authorization") String token){
+        if (!validateToken(token)) {
+            return null;
+        }
         return userDao.getUsers();
+    }
+
+
+    private boolean validateToken(String token){
+        String userId = jwtUtil.getKey(token);
+        return userId != null;
     }
 
     //Create a user
     @RequestMapping(value = "api/users",method = RequestMethod.POST)
     public void createUser(@RequestBody User user){
-
         Argon2 argon2 = Argon2Factory.create(Argon2Factory.Argon2Types.ARGON2id);
         String hash = argon2.hash(2,1024,1, user.getPassword());
         user.setPassword(hash);
         userDao.createUser(user);
-
-
-    }
-
-
-    //Update a user
-    @RequestMapping(value = "api/user/update", method = RequestMethod.PUT)
-    public User updateUser(){
-        User user = new User();
-        user.setName("Jonnathan");
-        user.setLastName("Campoberde");
-        user.setEmail("testing@testing.com");
-        user.setPhone("593984404457");
-        user.setPassword("testing");
-        return user;
     }
 
 
     //Delete a user
     @RequestMapping(value = "/api/users/delete/{id}", method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable Long id){
-        boolean result = userDao.deleteUser(id);
+    public void deleteUser(@RequestHeader(value="Authorization") String token,@PathVariable Long id){
+        if (!validateToken(token)) { return; }
+            userDao.deleteUser(id);
+
+        userDao.deleteUser(id);
 
     }
